requirements.txt

_libgcc_mutex==0.1
ca-certificates==2020.7.22
certifi==2020.6.20
cudatoolkit==10.1.243
libffi==3.2.1
openssl==1.0.2u
pip==20.0.2
python==3.6.3
setuptools==47.1.1
sqlite==3.31.1
tk==8.6.8
wheel==0.34.2
xz==5.2.5
zlib==1.2.11

absl-py==0.9.0
appdirs==1.4.4
astor==0.8.1
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
cachetools==4.1.0
chainer==7.7.0
chardet==3.0.4
cpython==0.0.6
cupy==7.7.0
cycler==0.10.0
cython==0.29.21
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
fastrlock==0.5
filelock==3.0.12
future==0.18.2
gast==0.3.3
google-auth==1.17.2
google-auth-oauthlib==0.4.1
grpcio==1.29.0
h5py==2.10.0
idna==2.9
imageio==2.8.0
importlib-metadata==1.6.1
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
jedi==0.17.0
jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
keras-applications==1.0.8
keras-preprocessing==1.1.2
kiwisolver==1.2.0
lmdb==0.98
mako==1.1.3
markdown==3.2.2
markupsafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
nose==1.3.7
notebook==6.0.3
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.2.0.34
packaging==20.4
pandas==1.1.0
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pillow==7.1.2
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.5.0
pycuda==2019.1.2
pygments==2.6.1
pymongo==3.11.0
pyparsing==2.4.7
pypi==2.1
pyrsistent==0.16.0
python-dateutil==2.8.1
pytools==2020.3
pywavelets==1.1.1
pyyaml==5.3.1
pyzmq==19.0.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.4.1
send2trash==1.5.0
six==1.15.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorboardx==2.0
tensorflow-gpu==1.12.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.6.3
torch==1.1.0
torchvision==0.3.0
tornado==6.0.4
traitlets==4.3.3
tqdm==4.49.0
typing-extensions==3.7.4.2
urllib3==1.25.9
wcwidth==0.2.4
webencodings==0.5.1
werkzeug==1.0.1
zipp==3.1.0

